Move functionality to GenTests
