--- src/test/resources/geninput/omitconstructor.txt
@@ -1,1 +1,0 @@
-.*<init>.*
