--- src/test/resources/geninput/omitmethod.txt
@@ -1,1 +1,0 @@
-.*method.*
